V1.3.2 - Various Fixes
Increased stack size of Crystal Hearts to 8.

Increased stack size of Crystal Grenades to 8.

Added better detailed death messages for heart/locust/grenade deaths.

Created a placeholder crystal planter that is the sole craftable planter. It better explains the crafting process for the variant blocks as a nondescript crystal planter.

Fixed an issue where crystal planters were not craftable.

Fixed an issue where obelisks could not be broken with chisels.